requirements.txt
# =============================================================================
# Financial Sentiment Analysis — Python Dependencies
# =============================================================================
# Install all dependencies with:
#   pip install -r requirements.txt
# =============================================================================

# --- Data handling ---
numpy
pandas

# --- Machine Learning ---
torch
transformers
scipy
xgboost                   # Gradient boosting for return prediction
scikit-learn              # ML utilities (metrics, preprocessing)

# --- Financial data ---
yfinance                  # Yahoo Finance API wrapper (stock prices)

# --- News data ---
requests                  # HTTP requests (Alpha Vantage news API)
feedparser                # RSS feed parsing for live news pipeline

# --- Environment management ---
python-dotenv             # Load API keys from .env file

# --- Dashboard ---
streamlit                 # Interactive web dashboard
plotly                    # Interactive charts for Streamlit

# --- Notebooks only (not needed for dashboard/src) ---
# Uncomment these if you plan to run the Jupyter notebooks:
# matplotlib
# seaborn
# textblob